Criando repositorio com Spring Data

+++ src/main/java/br/com/springboot/jdev_treinamento/model/Usuario.java
@@ -16,7 +16,7 @@ public class Usuario implements Serializable{
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seq_usuario")
-	private Long id;
+	private int id;
 	
 	private String nome;
 	
@@ -27,7 +27,7 @@ public class Usuario implements Serializable{
 	public long getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public String getNome() {
